Constants can now be defined as XML
Here's an example of two simple constants:

<abak name="simple values">
  <k ricef="ZTEST" field="BUKRS" context="A" value="1111"/>
  <k ricef="ZTEST" field="BUKRS" context="B" value="2222"/>
  <k ricef="ZTEST" field="UNAME" value="USER1"/>
</abak>

And here's an example of a range:

<abak name="a range">
  <k ricef="ZTEST" field="BUKRS">
    <v sign="I" option="EQ" low="1111"/>
    <v sign="I" option="BT" low="2000" high="3000"/>
  </k>
</abak>

Ideally either inline or complete <v> values should be allowed but
not both in one <k>. Unfortunately, due to not being able to understand
how the ST command <tt:cond> conditions work, both inline and <v>
values are allowed at the same time for one constant. This is not a
big problem but I'll still change it if I can figure this out.