LOCATION introduced
We intend to have multiple formats:

XML
JSON
CSV
etc

And the content for these can come from multiple locations:

inline
URL
server file
etc?

The current design requires one data source for each combination of the above format/location. This is not practical.

This commit introduces a new concept: location.

In the form of interface ZIF_ABAK_LOCATION which holds a string and has one implementation for each different location from where that stream can be fetched;

The data sources for each of the above formats receive an instance to a location and get their required parameters from there.